[feature] 캘린더에서 선택한 날짜의 전시회에 대한 기록 수정 구현 및 삭제 모달 위치 변경
